+++ test/util/copy_entity.js
@@ -10,8 +10,8 @@ test('copies a blank entity', t => {
         
         // let src = registry.createEntity();
         let src = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'alpha'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'alpha'}
         ]);
 
         let dst = copyEntity( registry, src );
@@ -27,8 +27,8 @@ test('copies a blank entity', t => {
 test('will retain dst components missing from the src by default', t => {
     return initialiseRegistry().then( registry => {
         
-        let src = registry.createEntity([  {id:'/component/position', x:2, y:-2}  ]);
-        let dst = registry.createEntity([  {id:'/component/name', 'name':'alpha'} ]);
+        let src = registry.createEntity([  {'@c':'/component/position', x:2, y:-2}  ]);
+        let dst = registry.createEntity([  {'@c':'/component/name', 'name':'alpha'} ]);
         let copy = copyEntity( registry, src, dst );
         
         t.equals( copy.Position.get('y'), -2 );
@@ -41,8 +41,8 @@ test('will retain dst components missing from the src by default', t => {
 test('will remove components missing from the src', t => {
     return initialiseRegistry().then( registry => {
         
-        let src = registry.createEntity([  {id:'/component/position', x:2, y:-2}  ]);
-        let dst = registry.createEntity([  {id:'/component/name', 'name':'alpha'} ]);
+        let src = registry.createEntity([  {'@c':'/component/position', x:2, y:-2}  ]);
+        let dst = registry.createEntity([  {'@c':'/component/name', 'name':'alpha'} ]);
         let copy = copyEntity( registry, src, dst, {delete:true} );
         
         t.equals( copy.Position.get('y'), -2 );
@@ -59,13 +59,13 @@ test('returns false if nothing changed', t => {
         
         // let src = registry.createEntity();
         let src = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'alpha'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'alpha'}
         ]);
 
         let dst = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'alpha'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'alpha'}
         ]);
 
         let [copy,hasChanged] = copyEntity( registry, src, dst, {returnChanged:true} );
@@ -83,13 +83,13 @@ test('returning whether anything was changed on the dst entity', t => {
         
         // let src = registry.createEntity();
         let src = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'alpha'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'alpha'}
         ]);
 
         let dst = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'beta'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'beta'}
         ]);
 
         let [copy,hasChanged] = copyEntity( registry, src, dst, {returnChanged:true} );
@@ -107,12 +107,12 @@ test('a missing dst component counts as a change', t => {
         
         // let src = registry.createEntity();
         let src = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
+            { '@c':'/component/position', x:2, y:-2 },
         ]);
 
         let dst = registry.createEntity([
-            { id:'/component/position', x:2, y:-2 },
-            { id:'/component/name', 'name':'alpha'}
+            { '@c':'/component/position', x:2, y:-2 },
+            { '@c':'/component/name', 'name':'alpha'}
         ]);
 
         let [copy,hasChanged] = copyEntity( registry, src, dst, {delete:true,returnChanged:true} );